requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch
torchvision
torchaudio

numpy
scipy
pandas
scikit-learn
matplotlib
seaborn
tqdm
joblib
threadpoolctl

jupyter
notebook
ipykernel
ipywidgets

pyyaml